pillar/hypervisor: fix discrepancy in setupSpec
kvm and xen sent an empty list of ioBundles to setupSpec, so that
it failed to lookup devices with `aa.LookupIoBundleAny(adapter.Name)`

turns out, we can just not call this piece of code in the case of
kvm and xen - only when using it directly from containerd; then
we also include the proper ioBundles
