+++ pkg/pillar/hypervisor/xen.go
@@ -128,10 +128,7 @@ func (ctx xenContext) Setup(status types.DomainStatus, config types.DomainConfig
 		return logError("failed to build domain config: %v", err)
 	}
 
-	spec, err := ctx.setupSpec(&status, &config, &types.AssignableAdapters{
-		Initialized:  false,
-		IoBundleList: []types.IoBundle{},
-	}, types.NewConfigItemValueMap(), status.OCIConfigDir)
+	spec, err := ctx.setupSpec(&status, &config, status.OCIConfigDir)
 
 	if err != nil {
 		return logError("failed to load OCI spec for domain %s: %v", status.DomainName, err)
